+++ src/OrchardCore/OrchardCore.DisplayManagement/Html/CssOrchardHelper.cs
@@ -2,9 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using OrchardCore.ContentManagement.Metadata.Models;
 using OrchardCore.DisplayManagement.Html;
-using OrchardCore.Mvc.Utilities;
 
 namespace OrchardCore;
 
@@ -86,39 +84,5 @@ public static class CssOrchardHelper
         }
 
         return additionalClasses.Concat(new[] { optionClasses });
-    }
-
-    private static IEnumerable<string> GetPartWrapperCssClasses(ContentTypePartDefinition partDefinition)
-    {
-        yield return "content-part-wrapper";
-        yield return $"content-part-wrapper-{partDefinition.PartDefinition.Name.HtmlClassify()}";
-
-        if (partDefinition.IsNamedPart())
-        {
-            yield return $"content-part-wrapper-{partDefinition.Name.HtmlClassify()}";
-        }
-    }
-
-    private static IEnumerable<string> GetFieldWrapperCssClasses(ContentPartFieldDefinition fieldDefinition)
-    {
-        yield return "field-wrapper";
-        yield return $"field-wrapper-{fieldDefinition.GetReusableFieldWrapperClassName().HtmlClassify()}";
-
-        if (fieldDefinition.IsNamedPart())
-        {
-            yield return $"field-wrapper-{fieldDefinition.GetFieldWrapperClassName().HtmlClassify()}";
-        }
-    }
-
-    public static string GetPartWrapperCssClasses(this IOrchardHelper helper, ContentTypePartDefinition partDefinition, bool skipThemeWrapperClasses = false)
-    {
-        var partClasses = GetPartWrapperCssClasses(partDefinition).ToArray();
-
-        return skipThemeWrapperClasses ? string.Join(' ', partClasses) : helper.GetWrapperCssClasses(partClasses);
-    }
-
-    public static string GetFieldWrapperCssClasses(this IOrchardHelper helper, ContentPartFieldDefinition fieldDefinition)
-    {
-        return helper.GetWrapperCssClasses(GetFieldWrapperCssClasses(fieldDefinition).ToArray());
     }
 }
